Define custom dagger scope and add some comments

+++ app/src/main/java/com/bernatgomez/apps/template/presenter/MainPresenter.java
@@ -37,7 +37,7 @@ public class MainPresenter extends BasePresenter {
     }
 
     @Subscribe
-    public void onDataReceived(Integer data) {
+    public void onDataReceived(Float data) {
         this.view.hideLoading();
     }
